requirements.txt
absl-py==2.3.0
agate==1.13.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
altgraph==0.17.4
amqp==5.3.1
anyio==4.9.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.10.0
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.3
binaryornot==0.4.4
blinker==1.9.0
build==1.2.2.post1
cachetools==6.1.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cookiecutter==2.6.0
cryptography==45.0.4
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
csvkit==2.1.0
cycler==0.12.1
daphne==4.2.1
datadispatch==1.0.0
DateTime==5.5
dbfread==2.0.7
distlib==0.4.0
dj-database-url==2.1.0
Django==4.2.16
django-allauth==65.9.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.11.2
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docutils==0.21.2
docx-mailmerge==0.5.0
easyocr==1.7.2
emails==0.6
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.41.1
greenlet==3.2.1
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
image==1.5.33
imageio==2.37.0
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.4
lazy_loader==0.4
leather==0.4.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
materialshapes==0.3
materialyoucolor==2.0.10
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==1.26.4
olefile==0.47
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
parsedatetime==2.6
pdf2image==1.17.0
pefile==2023.2.7
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.7
pluggy==1.5.0
premailer==3.10.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==4.25.8
psutil==6.1.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyMuPDF==1.25.5
PyMySQL==1.1.1
pyodbc==5.2.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.6.1
PyPDF2==3.0.1
pyperclip==1.9.0
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytesseract==0.3.13
pytz==2025.2
qrcode==8.2
redis==6.4.0
reportlab==4.4.0
requests==2.32.3
rich==13.9.4
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
service-identity==24.2.0
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
stripe==14.0.1
tasks==2.8.0
text-unidecode==1.3
tifffile==2025.5.10
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.1
torchvision==0.22.1
twilio==9.6.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.32.0
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.6.0
wikipedia==1.4.0
xgboost==3.0.2
xhtml2pdf==0.2.17
xlrd==2.0.2
xlsxwriter==3.2.5
yagmail==0.15.293
yarl==1.20.1
zipp==3.21
zope.interface==7.2